+++ descope/internal/mgmt/project_test.go
@@ -37,6 +37,32 @@ func TestProjectImport(t *testing.T) {
 	require.NoError(t, err)
 }
 
+func TestProjectImportWithExcludes(t *testing.T) {
+	mgmt := newTestMgmt(nil, helpers.DoOk(func(r *http.Request) {
+		req := map[string]any{}
+		require.NoError(t, helpers.ReadBody(r, &req))
+		excludes, ok := req["excludes"].([]any)
+		require.True(t, ok)
+		require.Equal(t, []any{"lists"}, excludes)
+	}))
+	req := &descope.ImportSnapshotRequest{
+		Files:    map[string]any{"foo": "bar"},
+		Excludes: []descope.SnapshotExclude{descope.SnapshotExcludeLists},
+	}
+	err := mgmt.Project().ImportSnapshot(context.Background(), req)
+	require.NoError(t, err)
+}
+
+func TestProjectImportWithInvalidExclude(t *testing.T) {
+	mgmt := newTestMgmt(nil, helpers.DoOk(nil))
+	req := &descope.ImportSnapshotRequest{
+		Files:    map[string]any{"foo": "bar"},
+		Excludes: []descope.SnapshotExclude{"pikachu"},
+	}
+	err := mgmt.Project().ImportSnapshot(context.Background(), req)
+	require.ErrorContains(t, err, "pikachu")
+}
+
 func TestValidateProjectImport(t *testing.T) {
 	resbody := map[string]any{
 		"ok":       false,