+++ descope/internal/mgmt/ssoapplication.go
@@ -91,6 +91,44 @@ func (s *ssoApplication) UpdateSAMLApplication(ctx context.Context, appRequest *
 	return err
 }
 
+func (s *ssoApplication) CreateWSFedApplication(ctx context.Context, appRequest *descope.WSFedApplicationRequest) (id string, err error) {
+	if appRequest == nil {
+		return "", utils.NewInvalidArgumentError("appRequest")
+	}
+	if appRequest.Name == "" {
+		return "", utils.NewInvalidArgumentError("appRequest.Name")
+	}
+
+	req := makeCreateUpdateWSFedApplicationRequest(appRequest)
+	httpRes, err := s.client.DoPostRequest(ctx, api.Routes.ManagementSSOApplicationWSFedCreate(), req, nil, "")
+	if err != nil {
+		return "", err
+	}
+	res := &struct {
+		ID string `json:"id"`
+	}{}
+	if err = utils.Unmarshal([]byte(httpRes.BodyStr), res); err != nil {
+		return "", err
+	}
+	return res.ID, nil
+}
+
+func (s *ssoApplication) UpdateWSFedApplication(ctx context.Context, appRequest *descope.WSFedApplicationRequest) error {
+	if appRequest == nil {
+		return utils.NewInvalidArgumentError("appRequest")
+	}
+	if appRequest.ID == "" {
+		return utils.NewInvalidArgumentError("appRequest.id")
+	}
+	if appRequest.Name == "" {
+		return utils.NewInvalidArgumentError("appRequest.Name")
+	}
+
+	req := makeCreateUpdateWSFedApplicationRequest(appRequest)
+	_, err := s.client.DoPostRequest(ctx, api.Routes.ManagementSSOApplicationWSFedUpdate(), req, nil, "")
+	return err
+}
+
 func (s *ssoApplication) Delete(ctx context.Context, id string) error {
 	if id == "" {
 		return utils.NewInvalidArgumentError("id")
@@ -138,25 +176,44 @@ func makeCreateUpdateOIDCApplicationRequest(appRequest *descope.OIDCApplicationR
 
 func makeCreateUpdateSAMLApplicationRequest(appRequest *descope.SAMLApplicationRequest) map[string]any {
 	return map[string]any{
+		"id":                        appRequest.ID,
+		"name":                      appRequest.Name,
+		"description":               appRequest.Description,
+		"enabled":                   appRequest.Enabled,
+		"logo":                      appRequest.Logo,
+		"loginPageUrl":              appRequest.LoginPageURL,
+		"useMetadataInfo":           appRequest.UseMetadataInfo,
+		"metadataUrl":               appRequest.MetadataURL,
+		"entityId":                  appRequest.EntityID,
+		"acsUrl":                    appRequest.AcsURL,
+		"certificate":               appRequest.Certificate,
+		"attributeMapping":          appRequest.AttributeMapping,
+		"groupsMapping":             appRequest.GroupsMapping,
+		"acsAllowedCallbacks":       appRequest.AcsAllowedCallbacks,
+		"defaultRelayState":         appRequest.DefaultRelayState,
+		"subjectNameIdType":         appRequest.SubjectNameIDType,
+		"subjectNameIdFormat":       appRequest.SubjectNameIDFormat,
+		"forceAuthentication":       appRequest.ForceAuthentication,
+		"logoutRedirectUrl":         appRequest.LogoutRedirectURL,
+		"defaultSignatureAlgorithm": appRequest.DefaultSignatureAlgorithm,
+	}
+}
+
+func makeCreateUpdateWSFedApplicationRequest(appRequest *descope.WSFedApplicationRequest) map[string]any {
+	return map[string]any{
 		"id":                  appRequest.ID,
 		"name":                appRequest.Name,
 		"description":         appRequest.Description,
 		"enabled":             appRequest.Enabled,
 		"logo":                appRequest.Logo,
 		"loginPageUrl":        appRequest.LoginPageURL,
-		"useMetadataInfo":     appRequest.UseMetadataInfo,
-		"metadataUrl":         appRequest.MetadataURL,
-		"entityId":            appRequest.EntityID,
-		"acsUrl":              appRequest.AcsURL,
-		"certificate":         appRequest.Certificate,
+		"realm":               appRequest.Realm,
+		"replyUrl":            appRequest.ReplyURL,
 		"attributeMapping":    appRequest.AttributeMapping,
 		"groupsMapping":       appRequest.GroupsMapping,
-		"acsAllowedCallbacks": appRequest.AcsAllowedCallbacks,
-		"defaultRelayState":   appRequest.DefaultRelayState,
-		"subjectNameIdType":   appRequest.SubjectNameIDType,
-		"subjectNameIdFormat": appRequest.SubjectNameIDFormat,
 		"forceAuthentication": appRequest.ForceAuthentication,
 		"logoutRedirectUrl":   appRequest.LogoutRedirectURL,
+		"errorRedirectUrl":    appRequest.ErrorRedirectURL,
 	}
 }